Move components into typography folder and capitalize component files
