+++ src/components/cards/cardProject/cardProject.tsx
@@ -1,23 +1,21 @@
 import React from 'react';
-import { Card } from '../../ui/container/card';
-import { Column } from '../../ui/container/column';
-import { Fill } from '../../ui/container/fill';
-import { Row } from '../../ui/container/row';
-import { Header } from '../../ui/content/header';
-import { Tag } from '../../ui/content/tag';
-import { UnstyledLink } from '../../ui/content/unstyledLink';
-import { MarginSmall } from '../../ui/margins/marginSmall';
+import { Card } from '../../ui/container/Card';
+import { Column } from '../../ui/container/Column';
+import { Fill } from '../../ui/container/Fill';
+import { Row } from '../../ui/container/Row';
+import { UnstyledLink } from '../../ui/content/UnstyledLink';
+import { MarginSmall } from '../../ui/margins/MarginSmall';
+import { Header } from '../../ui/typography/Header';
 import { animationSpring } from '../animations';
 
 interface CardProjectProps {
   title: string;
   summary: string;
-  tags?: string[];
   urlSource: string;
   url?: string;
 }
 
-export const CardProject: React.FC<CardProjectProps> = ({ title, summary, tags, urlSource, url }) => {
+export const CardProject: React.FC<CardProjectProps> = ({ title, summary, urlSource, url }) => {
   return (
     <Card {...animationSpring}>
       <Row style={{ height: '100%' }}>
@@ -33,7 +31,6 @@ export const CardProject: React.FC<CardProjectProps> = ({ title, summary, tags, 
           </UnstyledLink>
           <Fill />
           <Row>
-            {tags && tags.map((tag, index) => <Tag key={tag + index}>{tag}</Tag>)}
             <Fill />
             {url && <a href={url}>Visa</a>}
           </Row>
