+++ src/components/navbar/navbar.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useDarkMode } from '../../hooks/useDarkMode';
-import { Fill } from '../ui/container/fill';
-import { ResponsiveContainer } from '../ui/container/responsiveContainer';
-import { Row } from '../ui/container/row';
-import { UnstyledInternalLink } from '../ui/content/unstyledLink';
-import { MarginSmall } from '../ui/margins/marginSmall';
-import { DarkModeToggleIcons } from './darkModeToggleIcons';
+import { Fill } from '../ui/container/Fill';
+import { ResponsiveContainer } from '../ui/container/ResponsiveContainer';
+import { Row } from '../ui/container/Row';
+import { UnstyledInternalLink } from '../ui/content/UnstyledLink';
+import { MarginSmall } from '../ui/margins/MarginSmall';
+import { DarkModeToggleIcons } from './DarkModeToggleIcons';
 import { NavbarContentWrapper, NavbarItem, NavbarWrapper } from './nav.styles';
 
 export const Navbar = () => {
