--- src/components/ui/content/tag.tsx
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-export const Tag = styled.span`
-  color: white;
-  background: var(--text-muted-color);
-  border-radius: var(--border-radius-small);
-  padding: 4px 8px;
-  margin: 2px;
-  font-size: 0.8rem;
-`;
-
-Tag.displayName = 'Tag';
