+++ src/components/ui/margins/marginLarge.tsx
@@ -1,4 +1,4 @@
-import { Margin } from './margin';
+import { Margin } from './Margin';
 
 export const MarginLarge: React.FC = () => <Margin size={2} />;
