+++ src/components/ui/margins/marginMedium.tsx
@@ -1,4 +1,4 @@
-import { Margin } from './margin';
+import { Margin } from './Margin';
 
 export const MarginMedium: React.FC = () => <Margin size={1} />;
