+++ src/components/ui/margins/marginSmall.tsx
@@ -1,4 +1,4 @@
-import { Margin } from './margin';
+import { Margin } from './Margin';
 
 export const MarginSmall: React.FC = () => <Margin size={0} />;
