+++ src/pages/404.tsx
@@ -1,11 +1,11 @@
-import { Seo } from '../components/seo/seo';
-import { ResponsiveContainer } from '../components/ui/container/responsiveContainer';
-import { MarginLarge } from '../components/ui/margins/marginLarge';
-import { TextCenter } from '../components/ui/content/textCenter';
+import { Seo } from '../components/seo/Seo';
 import { ButtonPrimary } from '../components/ui/buttons/ButtonPrimary';
-import { MarginMedium } from '../components/ui/margins/marginMedium';
-import { Column } from '../components/ui/container/column';
-import { UnstyledInternalLink } from '../components/ui/content/unstyledLink';
+import { Column } from '../components/ui/container/Column';
+import { ResponsiveContainer } from '../components/ui/container/ResponsiveContainer';
+import { UnstyledInternalLink } from '../components/ui/content/UnstyledLink';
+import { MarginLarge } from '../components/ui/margins/MarginLarge';
+import { MarginMedium } from '../components/ui/margins/MarginMedium';
+import { TextCenter } from '../components/ui/typography/TextCenter';
 
 const page404 = () => {
   const title = '404! Sidan kunde inte hittas';
