+++ src/pages/_app.tsx
@@ -4,11 +4,11 @@ import { AppProps } from 'next/app';
 import Head from 'next/head';
 import * as React from 'react';
 import SEO from '../../next-seo.config';
-import { Footer } from '../components/footer/footer';
-import { mdxComponents } from '../components/mdx/mdxComponents';
+import { Footer } from '../components/footer/Footer';
+import { mdxComponents } from '../components/mdx/MdxComponents';
 import { NavAdjustedContent } from '../components/navbar/nav.styles';
-import { Navbar } from '../components/navbar/navbar';
-import { MarginLarge } from '../components/ui/margins/marginLarge';
+import { Navbar } from '../components/navbar/Navbar';
+import { MarginLarge } from '../components/ui/margins/MarginLarge';
 import { ContentProvider } from '../context/ContentContext';
 import { ThemeProvider } from '../context/ThemeContext';
 import { getAllBlogPosts } from '../data/blogPosts';
