+++ src/pages/blogg/hur-aftonbladet-pajade-min-sajt-och-varfor-det-ar-okej-att-lata-det-ligga-i-backloggen-tillsvidare.mdx
@@ -7,7 +7,7 @@ date: 2021-07-09
 
 ## Sajten
 
-Jag har tidigare [skrivit lite](https://www.danielvernberg.se/blogg/att-vacka-liv-i-ett-3-ar-gammalt-projekt) om en sajt/scraper jag byggt - som via en api-endpoint parsa startsidan på Aftonbladet och sparar ned viss data om Aftonbladets startsida i Google Firebase.
+Jag har tidigare [skrivit lite](/blogg/att-vacka-liv-i-ett-3-ar-gammalt-projekt) om en sajt/scraper jag byggt - som via en api-endpoint parsa startsidan på Aftonbladet och sparar ned viss data om Aftonbladets startsida i Google Firebase.
 
 Endpointen anropas dagligen klockan 12 och har nu fungerat felfritt i ett antal år, fram till i år.
